# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: CareLink API
  version: 0.0.1
  description: >-
    Caregiver slice 1 is implemented for GET /caregivers/me, GET /caregivers/me/schedule,
    and GET /visits/{visitId}/work-pack; these definitions supersede the matching draft operations.
    Includes the existing Auth contract and the specified FM01-FM05 family contracts. Family
    operations remain x-status: draft until implemented; inclusion in Final does not imply a working
    backend. Shared paths retain their non-family response alternatives. The family owner supplies
    only the family use-case behaviour; other use cases, binding management and platform
    implementation are outside this change.
  license:
    name: UNLICENSED
servers:
  - url: /api
    description: Same-origin, relative to whichever host serves the front end.
...
paths:
...
  /incidents:
    post:
      tags: [Incidents]
      summary: Submit a caregiver incident report (CG04)
      operationId: reportCaregiverIncident
      x-status: implemented
      x-implemented-roles: [CAREGIVER]
      security: [{sessionCookie: []}]
      description: >-
        Requires CSRF. Actor and elder are derived from the authenticated caregiver and current visit assignment.
        Open work pauses to EXCEPTION; terminal visits retain their state. SCHEDULED reports require the start
        time to have been reached. A UUID identifies one immutable command; equal retries return the original
        report without new notifications. Receipt status is read live; visitVersion identifies the original write.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CaregiverIncidentCommand'}
      responses:
        '201':
          description: Report created and routed through MG05 in the same transaction.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CaregiverIncidentResult'}
        '200':
          description: An identical successful command replayed; no additional business effects.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CaregiverIncidentResult'}
        '400': {description: Invalid request body.}
        '401': {description: Authentication required.}
        '403': {description: 'Wrong role, CSRF failure or no longer assigned.'}
        '404': {description: Visit not found.}
        '409': {description: State/version/command conflict; refresh before creating a different command.}
        '503': {description: Storage unavailable; reconcile saved results before retrying the same command.}
  /caregivers/me/incidents:
    get:
      tags: [Caregivers, Incidents]
      summary: Read reports submitted by this caregiver account
      operationId: getOwnCaregiverIncidentReports
      x-status: implemented
      security: [{sessionCookie: []}]
      description: Only source CAREGIVER and reportedByUserId equal to the session account; historical ownership survives visit reassignment. Sorted reportedAt/id descending.
      parameters:
        - {name: visitId, in: query, schema: {type: integer, format: int64, minimum: 1}}
        - {name: page, in: query, schema: {type: integer, minimum: 0, default: 0}}
        - {name: size, in: query, schema: {type: integer, minimum: 1, maximum: 50, default: 20}}
      responses:
        '200':
          description: Safe own-report page, never a manager queue.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: {type: array, items: {$ref: '#/components/schemas/CaregiverIncidentView'}}
                  page: {type: integer}
                  size: {type: integer}
                  totalElements: {type: integer, format: int64}
        '400': {description: Invalid page or visit filter.}
        '401': {description: Authentication required.}
        '403': {description: Enabled caregiver account required.}
  /caregivers/me/incidents/{id}:
    get:
      tags: [Caregivers, Incidents]
      summary: Read a safe receipt for an incident this account reported
      operationId: getOwnCaregiverIncidentReport
      x-status: implemented
      security: [{sessionCookie: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer, format: int64}}
      responses:
        '200':
          description: Own report with live incident status, no responder identities or internal logs.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CaregiverIncidentView'}
        '401': {description: Authentication required.}
        '403': {description: Enabled caregiver account required.}
        '404': {description: Missing or not submitted by this account; same unavailable response.}
...
  /caregivers/{caregiverId}:
    parameters:
      - $ref: '#/components/parameters/CaregiverId'
    get:
      tags:
        - Caregivers
      summary: Get a caregiver profile
      x-status: draft
      x-family-status: implemented
      responses:
        '200':
          description: >-
            The implemented FAMILY response contains only FamilyCaregiverView fields. The
            original Caregiver response is retained for other roles, whose implementation is
            pending. The server selects the projection from the authenticated role.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/FamilyCaregiverView'
                  - $ref: '#/components/schemas/Caregiver'
              examples:
                familyProfile:
                  summary: Public profile of a caregiver with an authorized visit relationship
                  value:
                    id: 201
                    fullName: Lim Jia Hui
                    dialects:
                      - Mandarin
                      - Hokkien
                noDialects:
                  summary: Public profile with no recorded dialects
                  value:
                    id: 202
                    fullName: Daniel Goh
                    dialects: []
        '400':
          description: The caregiverId path value is not a signed 64-bit integer.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: >-
            No valid authenticated session. Shared security determines the response body;
            clients must handle the HTTP status without requiring a JSON body.
        '403':
          description: >-
            The caller lacks FAMILY, the current family account/profile is unavailable, or
            caregiverId has no visit relationship to a currently readable elder. This includes
            unrelated and nonexistent integer IDs, including zero and negative IDs. The
            relationship check precedes the profile lookup.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: >-
            The caregiver profile cannot be found after the family's visit relationship has
            passed authorization. An ID without an authorized relationship returns 403 first.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '503':
          $ref: '#/components/responses/FamilyAuditUnavailable'
      x-contract-status: specified
      x-family-use-cases:
        - UC-FM02
      x-family-owner: Member 4 (FM01-FM05 behaviour); source data remains with its assigned owner.
      security:
        - sessionCookie: []
      operationId: getCaregiver
      description: >-
        The implemented FAMILY detail reads the public profile of a caregiver assigned to a
        current or historical visit of an elder the caller can currently access. The visit's
        caregiverId must match; a visit_assignment replacement-history row alone grants no
        access. This relationship is independent of the selected schedule week or visit status.
        Historical visits remain eligible only while the elder binding remains readable.

        Identity comes from the authenticated session; the account must currently be enabled,
        hold FAMILY and have a family profile. Each request checks ACTIVE, unexpired bindings
        with FULL or READ_ONLY scope. Expiry is evaluated in Asia/Singapore and access ends
        exactly at expiresAt. Revocation or expiry takes effect on subsequent requests. The
        client cannot select a family identity or response projection. This read-only GET
        requires a valid session but no CSRF header.

        Example: GET /api/caregivers/201. The path value must fit a signed 64-bit integer;
        malformed or overflowing values return 400. Authorization runs before the profile
        lookup: no eligible visit relationship returns 403, including an unrelated or
        nonexistent ID. Only a missing profile after successful authorization returns 404.

        The response contains only id, fullName and dialects. Stored comma-separated dialects
        become an array: trim each item and omit empty items while preserving order, spelling
        and duplicates. Null or blank source values become []. No phone, user/account ID,
        sector, internal status, performance or credential details are returned. An inactive
        caregiver's public profile remains readable through an eligible historical visit.

        Other-role profile contracts remain planned and their original response schema is
        retained. Their implementation is pending; non-FAMILY callers currently receive 403.

        Validated requests reaching the FAMILY application query persist a READ record with
        the actor app_user ID (when resolvable), resource and outcome: OK, including empty
        results; DENIED for access refusal; or FAILED for lookup/query errors. Records
        commit independently with server occurred_at in Asia/Singapore. Authentication 401,
        method-role rejection 403 and MVC validation/conversion 400 rejected before the
        FAMILY application query are outside this audit. Existing request logging is
        unchanged. Audit storage failure returns 503 without protected data. Records
        exclude response bodies, raw query strings, exception details and secrets. Reading
        does not acknowledge notifications or incidents. The resource is CAREGIVER,
        identified by the requested caregiverId.
      x-family-response-schema:
        $ref: '#/components/schemas/FamilyCaregiverView'
...
  /incidents/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags:
        - Incidents
      summary: Get an incident and its full timeline
      x-status: draft
      description: >-
        Who reported it, who responded, every notification attempt, and when the bound family member
        acknowledged it — this is the query that replaces the three-day chat-log reconstruction.

        Family reads use GET /family/incidents/{id}; this endpoint remains MANAGER-only.
      responses:
        '200':
          description: The manager incident detail and full handling timeline.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentDetail'
        '401':
          $ref: '#/components/responses/FamilyUnauthorized'
        '403':
          $ref: '#/components/responses/FamilyForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      x-contract-status: specified
      security:
        - sessionCookie: []
      operationId: getIncident
  /family/incidents/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags:
        - Incidents
      summary: Get the current family's safe incident detail
      operationId: getFamilyIncident
      x-status: implemented
      x-implemented-roles: [FAMILY]
      x-family-status: implemented
      x-contract-status: specified
      x-family-use-cases: [UC-FM05]
      x-family-owner: Member 4 (FM01-FM05 behaviour); source data remains with its assigned owner.
      security:
        - sessionCookie: []
      description: >-
        UC-FM05. Resolve the current family profile from an enabled account with the current
        FAMILY role. Require an ACTIVE, unexpired binding to the actual incident elder.
        FULL and READ_ONLY both permit this read. Recheck authorization on every request;
        client familyMemberId, elderId or projection parameters cannot select another identity
        or override the incident's elder. A notification is not required for access.
        Return safe incident facts, global handling status and only this family's receipt;
        exclude raw coordinates, location text, staff account IDs, playbook, internal timeline
        and other family receipts. Times include the Singapore +08:00 offset.
        A missing receipt returns an empty projection with the current familyMemberId and
        null id, viewedAt, acknowledgedAt and responseNote. GET never inserts or updates a
        receipt, marks a notification read or changes incident handling. Record the account,
        incident, outcome and server access time in the internal audit; an audit outage
        returns 503 without care content. An absent incident returns 404 and inaccessible
        incident returns 403. The manager's original GET /incidents/{id} remains separate.
        acknowledgeBy is this family's immutable response deadline established by the FM05
        Observer's first successful urgent IN_APP message creation (default two hours).
        It is null without an FM05 window, even when a legacy notification exists; never copy
        manager respondBy. The consumer is implemented; production event hookup and old family
        generator handoff remain pending. POST /incidents/{id}/view records a successful display;
        POST /incidents/{id}/acknowledge records explicit acknowledgement. Neither marks a notification read.
      responses:
        '200':
          description: Safe incident detail and only the current family's acknowledgement.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FamilyIncidentDetail'
        '400':
          $ref: '#/components/responses/FamilyValidationError'
        '401':
          $ref: '#/components/responses/FamilyUnauthorized'
        '403':
          $ref: '#/components/responses/FamilyForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/FamilyAuditUnavailable'
      x-family-response-schema:
        $ref: '#/components/schemas/FamilyIncidentDetail'
...
components:
  schemas:
...
    CaregiverCommandIdentity:
      type: object
      required: [expectedVersion, clientRequestId]
      properties:
        expectedVersion: {type: integer, minimum: 0}
        clientRequestId: {type: string, format: uuid, description: Immutable command key; retain original payload and key when the result is unknown.}
    CaregiverIncidentCommand:
      allOf:
        - $ref: '#/components/schemas/CaregiverCommandIdentity'
        - type: object
          required: [visitId, category, severity, description]
          properties:
            visitId: {type: integer, format: int64, minimum: 1}
            category: {type: string, enum: [SOS, MEDICAL, FALL, SERVICE, OTHER]}
            severity: {type: string, enum: [LOW, MEDIUM, HIGH]}
            description: {type: string, minLength: 1, maxLength: 2000, description: Observed facts and actions; no diagnosis.}
    CaregiverIncidentReport:
      type: object
      required: [id, visitId, category, severity, description, status, reportedAt]
      properties:
        id: {type: integer, format: int64}
        visitId: {type: integer, format: int64}
        category: {type: string, enum: [SOS, MEDICAL, FALL, SERVICE, OTHER]}
        severity: {type: string, enum: [LOW, MEDIUM, HIGH]}
        description: {type: string}
        status: {type: string, enum: [OPEN, ACKNOWLEDGED, IN_PROGRESS, RESOLVED, UNRESOLVED_ESCALATED]}
        reportedAt: {type: string, description: Singapore local date-time}
        respondBy: {type: string, nullable: true, description: Singapore local date-time; a deadline is not confirmation of take-over.}
        resolvedAt: {type: string, nullable: true, description: Singapore local date-time}
    CaregiverIncidentView:
      type: object
      required: [report, clientRequestId]
      properties:
        report: {$ref: '#/components/schemas/CaregiverIncidentReport'}
        clientRequestId: {type: string, format: uuid, nullable: true}
    CaregiverIncidentResult:
      allOf:
        - $ref: '#/components/schemas/CaregiverIncidentView'
        - type: object
          required: [visitVersion, replayed]
          properties:
            visitVersion: {type: integer}
            replayed: {type: boolean}
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string

    # Family projections and referenced shared schemas.
...
    FamilyProblemDetail:
      type: object
      required:
        - type
        - title
        - status
      description: >-
        Error contract for FM01-FM05. Optional code is a stable machine-readable reason; fields contains
        input validation messages.
      properties:
        type:
          type: string
          format: uri
          example: about:blank
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
      additionalProperties: true
      example:
        type: about:blank
        title: Insufficient permission
        status: 403
        detail: The current family member cannot access this resource.
...
    FamilyCaregiverView:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        fullName:
          type: string
          minLength: 1
        dialects:
          type: array
          description: >-
            Dialects from the source comma-separated string. Items are trimmed and empty
            items omitted; order, spelling and duplicates are preserved. Null or blank
            source values produce an empty array.
          items:
            type: string
      description: >-
        Public profile only: id, fullName and dialects. No phone, user/account ID, sector,
        private contact details, staff performance, internal availability state or credentials.
      required:
        - id
        - fullName
        - dialects
    Caregiver:
      type: object
      required:
        - id
        - fullName
        - status
      properties:
        id:
          type: integer
          format: int64
        fullName:
          type: string
        phone:
          type: string
        sector:
          type: string
        dialects:
          type: array
          items:
            type: string
        status:
          $ref: '#/components/schemas/CaregiverStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CaregiverStatus:
      type: string
      enum:
      - ONBOARDING
      - AVAILABLE
      - BUSY
      - INACTIVE
...
    FamilyIncidentDetail:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
        elderId:
          type: integer
          format: int64
        visitId:
          type: integer
          format: int64
          nullable: true
        source:
          $ref: '#/components/schemas/IncidentSource'
        category:
          $ref: '#/components/schemas/IncidentCategory'
        severity:
          $ref: '#/components/schemas/IncidentSeverity'
        status:
          $ref: '#/components/schemas/IncidentStatus'
        description:
          type: string
          nullable: true
        reportedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        acknowledgeBy:
          type: string
          format: date-time
          nullable: true
          description: >-
            Null without an FM05 personal window; legacy notifications do not establish one.
            The family member response deadline: first successfully created urgent IN_APP
            notification for this incident/family member plus the configured window (default two hours). Null when no
            response window was established, including when notification creation has not succeeded;
            null does not prevent an authorised read, view or acknowledgement. Stable across polling,
            email retries and further notifications for the same incident. Never the manager respondBy
            deadline. Later notification creation does not clear an existing acknowledgement or reopen
            waiting reminders.
        acknowledgement:
          $ref: '#/components/schemas/FamilyIncidentAcknowledgement'
      description: >-
        Safe incident facts and global handling status for the currently authorised family member.
        Notification existence or delivery is not an access prerequisite. No raw location coordinates,
        playbook, internal log, staff account IDs or other family acknowledgements.
      required:
        - id
        - elderId
        - visitId
        - source
        - category
        - severity
        - status
        - description
        - reportedAt
        - resolvedAt
        - acknowledgeBy
        - acknowledgement
    IncidentSource:
      type: string
      description: >-
        SYSTEM_MISSED_CHECKIN includes unassigned/leave-related system events and SYS03 assigned-but-overdue
        check-in alerts. SYS03 has an independent once-per-Visit durable trigger, defaults to strictly more than
        ten minutes after scheduledStart within a 24-hour lookback, and does not change Visit status.
      enum:
        - CAREGIVER
        - ELDER_SOS
        - ELDER_SERVICE_DISPUTE
        - SYSTEM_MISSED_CHECKIN
    IncidentCategory:
      type: string
      enum:
        - SOS
        - MEDICAL
        - FALL
        - SERVICE
        - OTHER
    IncidentSeverity:
      type: string
      enum:
        - LOW
        - MEDIUM
        - HIGH
    IncidentStatus:
      type: string
      enum:
        - OPEN
        - ACKNOWLEDGED
        - IN_PROGRESS
        - RESOLVED
        - UNRESOLVED_ESCALATED
    FamilyIncidentAcknowledgement:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          nullable: true
          description: >-
            Null in a detail read before a view/acknowledgement record is created. Non-null after a
            successful view or acknowledge command.
        incidentId:
          type: integer
          format: int64
          minimum: 1
        familyMemberId:
          type: integer
          format: int64
          minimum: 1
        viewedAt:
          type: string
          format: date-time
          nullable: true
        acknowledgedAt:
          type: string
          format: date-time
          nullable: true
        responseNote:
          type: string
          maxLength: 255
          nullable: true
      description: >-
        Current family member only. Viewed and acknowledged timestamps are independent and preserve
        their first server-recorded values.
      required:
        - id
        - incidentId
        - familyMemberId
        - viewedAt
        - acknowledgedAt
        - responseNote
    IncidentDetail:
      allOf:
      - $ref: '#/components/schemas/Incident'
      - type: object
        properties:
          playbookCode:
            type: string
            nullable: true
          log:
            type: array
            items:
              $ref: '#/components/schemas/IncidentLogEntry'
          acknowledgements:
            type: array
            items:
              $ref: '#/components/schemas/IncidentAcknowledgement'
          timeline:
            type: array
            description: Handling history for the incident.
            items:
              $ref: '#/components/schemas/IncidentLogEntry'
    Incident:
      type: object
      required:
        - id
        - elderId
        - source
        - severity
        - status
      properties:
        id:
          type: integer
          format: int64
        elderId:
          type: integer
          format: int64
        visitId:
          type: integer
          format: int64
          nullable: true
        reportedByUserId:
          type: integer
          format: int64
          nullable: true
        responderUserId:
          type: integer
          format: int64
          nullable: true
        source:
          $ref: '#/components/schemas/IncidentSource'
        category:
          $ref: '#/components/schemas/IncidentCategory'
        severity:
          $ref: '#/components/schemas/IncidentSeverity'
        status:
          $ref: '#/components/schemas/IncidentStatus'
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
        locationText:
          type: string
          nullable: true
        description:
          type: string
        respondBy:
          type: string
          format: date-time
          nullable: true
          description: Countdown deadline; past this, the scheduled job escalates automatically.
        reportedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
    IncidentLogEntry:
      type: object
      properties:
        id:
          type: integer
          format: int64
        actor:
          type: string
        action:
          type: string
        detail:
          type: string
        occurredAt:
          type: string
          format: date-time
    IncidentAcknowledgement:
      type: object
      properties:
        id:
          type: integer
          format: int64
        incidentId:
          type: integer
          format: int64
        familyMemberId:
          type: integer
          format: int64
        viewedAt:
          type: string
          format: date-time
          nullable: true
        acknowledgedAt:
          type: string
          format: date-time
          nullable: true
        responseNote:
          type: string
          nullable: true
...
  parameters:
...
    CaregiverId:
      name: caregiverId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    FamilyValidationError:
      description: >-
        Invalid input, date range, week start, unexpected request field, or unsupported family filter.
        Optional fields maps field names to validation messages.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/FamilyProblemDetail'
    FamilyUnauthorized:
      description: No valid authenticated session.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/FamilyProblemDetail'
    FamilyForbidden:
      description: >-
        The current caller lacks the family role, resource ownership, recipient identity, or a
        currently active, unexpired binding/access scope. An authenticated caller does not gain access
        by changing an ID.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/FamilyProblemDetail'
    NotFound:
      description: No resource with this id.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    FamilyAuditUnavailable:
      description: >-
        The family access record could not be committed. After any query outcome, failure to
        record access returns 503 instead of the result or original error, without protected data.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            title: Access recording unavailable
            status: 503
            detail: Access could not be recorded. Please retry later.
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: >
        Server-side session established by `POST /auth/login`. State-changing requests must also echo
        the `XSRF-TOKEN` cookie value in an `X-XSRF-TOKEN` header.